[HttpClient] add TimeoutExceptionInterface
