+++ src/Symfony/Component/HttpClient/Chunk/ErrorChunk.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\HttpClient\Chunk;
 
+use Symfony\Component\HttpClient\Exception\TimeoutException;
 use Symfony\Component\HttpClient\Exception\TransportException;
 use Symfony\Contracts\HttpClient\ChunkInterface;
 
@@ -61,7 +62,7 @@ class ErrorChunk implements ChunkInterface
     public function isFirst(): bool
     {
         $this->didThrow = true;
-        throw new TransportException($this->errorMessage, 0, $this->error);
+        throw null !== $this->error ? new TransportException($this->errorMessage, 0, $this->error) : new TimeoutException($this->errorMessage);
     }
 
     /**
@@ -70,7 +71,7 @@ class ErrorChunk implements ChunkInterface
     public function isLast(): bool
     {
         $this->didThrow = true;
-        throw new TransportException($this->errorMessage, 0, $this->error);
+        throw null !== $this->error ? new TransportException($this->errorMessage, 0, $this->error) : new TimeoutException($this->errorMessage);
     }
 
     /**
@@ -79,7 +80,7 @@ class ErrorChunk implements ChunkInterface
     public function getInformationalStatus(): ?array
     {
         $this->didThrow = true;
-        throw new TransportException($this->errorMessage, 0, $this->error);
+        throw null !== $this->error ? new TransportException($this->errorMessage, 0, $this->error) : new TimeoutException($this->errorMessage);
     }
 
     /**
@@ -88,7 +89,7 @@ class ErrorChunk implements ChunkInterface
     public function getContent(): string
     {
         $this->didThrow = true;
-        throw new TransportException($this->errorMessage, 0, $this->error);
+        throw null !== $this->error ? new TransportException($this->errorMessage, 0, $this->error) : new TimeoutException($this->errorMessage);
     }
 
     /**
@@ -119,7 +120,7 @@ class ErrorChunk implements ChunkInterface
     {
         if (!$this->didThrow) {
             $this->didThrow = true;
-            throw new TransportException($this->errorMessage, 0, $this->error);
+            throw null !== $this->error ? new TransportException($this->errorMessage, 0, $this->error) : new TimeoutException($this->errorMessage);
         }
     }
 }
